salary changes

+++ Backend/TenantManagementSystem/SalaryRecordRequest.cs
@@ -5,7 +5,7 @@
         public int EmployeeId { get; set; } // Assuming EmployeeId is provided
         public string SalaryMonth { get; set; }
         public int Salary { get; set; }
-        public int Leaves { get; set; }
+        public decimal Leaves { get; set; }
         public int Deductions { get; set; }
         public int NetPay { get; set; }
     }
